changed from double to float. tests still unchaged. need to doclear

+++ src/functions/activations.cpp
@@ -3,12 +3,12 @@
 
 LPP::Activation::~Activation() {}
 
-double LPP::Identity::apply_activation(double x) const
+float LPP::Identity::apply_activation(float x) const
 {
     return x;
 }
 
-double LPP::Identity::apply_derivative(double x) const
+float LPP::Identity::apply_derivative(float x) const
 {
     return 1.0;
 }
@@ -17,12 +17,12 @@ const std::string& LPP::Identity::who() const {
     return STRING_IDENTITY;
 }
 
-double LPP::ReLU::apply_activation(double x) const
+float LPP::ReLU::apply_activation(float x) const
 {
     return (x >= 0) ? x : 0.0;
 }
 
-double LPP::ReLU::apply_derivative(double x) const 
+float LPP::ReLU::apply_derivative(float x) const 
 {
     // Mathematically speaking, it should be undefined at zero
     // For convenience, we assign it 1
@@ -33,14 +33,14 @@ const std::string& LPP::ReLU::who() const {
     return STRING_RELU;
 }
 
-double LPP::Sigmoid::apply_activation(double x) const
+float LPP::Sigmoid::apply_activation(float x) const
 {
     return 1 / (1 + std::exp(-x));
 }
 
-double LPP::Sigmoid::apply_derivative(double x) const
+float LPP::Sigmoid::apply_derivative(float x) const
 {
-    const double s = LPP::Sigmoid::apply_activation(x);
+    const float s = LPP::Sigmoid::apply_activation(x);
     return s * (1-s);
 }
 
@@ -48,14 +48,14 @@ const std::string& LPP::Sigmoid::who() const {
     return STRING_SIGMOID;
 }
 
-double LPP::Tanh::apply_activation(double x) const
+float LPP::Tanh::apply_activation(float x) const
 {
     return std::tanh(x);
 }
 
-double LPP::Tanh::apply_derivative(double x) const
+float LPP::Tanh::apply_derivative(float x) const
 {
-    const double s = std::cosh(x);
+    const float s = std::cosh(x);
     return 1 / (s * s);
 }
